requirements.txt
boto3==1.17.88
botocore==1.20.97
click==7.1.2
Flask==1.1.4
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==2.0.1
nodeenv==1.6.0
Pillow==8.2.0
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.5
Werkzeug==1.0.1
